+++ src/main/java/de/johni0702/minecraft/bobby/commands/MergeWorldsCommand.java
@@ -1,0 +1,39 @@
+package de.johni0702.minecraft.bobby.commands;
+
+import com.mojang.brigadier.Command;
+import com.mojang.brigadier.context.CommandContext;
+import de.johni0702.minecraft.bobby.FakeChunkManager;
+import de.johni0702.minecraft.bobby.Worlds;
+import de.johni0702.minecraft.bobby.ext.ClientChunkManagerExt;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+import net.minecraft.client.world.ClientWorld;
+import net.minecraft.text.Text;
+
+import static com.mojang.brigadier.arguments.IntegerArgumentType.getInteger;
+
+public class MergeWorldsCommand implements Command<FabricClientCommandSource> {
+    @Override
+    public int run(CommandContext<FabricClientCommandSource> context) {
+        FabricClientCommandSource source = context.getSource();
+        ClientWorld world = source.getWorld();
+
+        FakeChunkManager bobbyChunkManager = ((ClientChunkManagerExt) world.getChunkManager()).bobby_getFakeChunkManager();
+        if (bobbyChunkManager == null) {
+            source.sendError(Text.translatable("bobby.upgrade.not_enabled"));
+            return 0;
+        }
+
+        Worlds worlds = bobbyChunkManager.getWorlds();
+        if (worlds == null) {
+            source.sendError(Text.translatable("bobby.dynamic_multi_world.not_enabled"));
+            return 0;
+        }
+
+        int sourceId = getInteger(context, "source");
+        int targetId = getInteger(context, "target");
+
+        worlds.userRequestedMerge(source, sourceId, targetId);
+
+        return 0;
+    }
+}