+++ src/main/java/de/johni0702/minecraft/bobby/util/RegionPos.java
@@ -28,4 +28,17 @@ public record RegionPos(int x, int z) {
     public static RegionPos from(ChunkPos chunkPos) {
         return new RegionPos(chunkPos.getRegionX(), chunkPos.getRegionZ());
     }
+
+    @Override
+    public int hashCode() {
+        return x ^ Integer.rotateRight(z, 16);
+    }
+
+    // The standard `Long.hashCode` results in close to worst-case performance with packet coordinates because it
+    // simply boils down to `x ^ z` and so 7/0 has the same hash code (7) as 6/1, 5/2, 4/3, 3/4, 2/5, 1/6 and 0/7.
+    public static int hashCode(long coords) {
+        int x = (int) (coords << 32 >> 32);
+        int z = (int) (coords >> 32);
+        return x ^ Integer.rotateRight(z, 16);
+    }
 }